requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.2.2
pinecone-client==2.2.4
python-dotenv==1.0.0
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.12.1
psycopg2-binary==2.9.9 